+++ src/main/java/com/hbm/main/MainRegistry.java
@@ -1054,6 +1054,7 @@ public class MainRegistry {
 		event.registerServerCommand(new CommandRadiation());
 		event.registerServerCommand(new CommandStations());
 		event.registerServerCommand(new CommandEnableNukes());
+		event.registerServerCommand(new CommandEnableRedstoneGroundNukes());
 		event.registerServerCommand(new CommandBedrockDrop());
 		MinecraftForge.EVENT_BUS.register(new NukeScheduler());
 		FMLCommonHandler.instance().bus().register(new NukeScheduler());